Spack-solvable versions, add tests to Spack workflow (#2156)
## Changes

1. Spack interprets version info consisting of 2-tuples as ranges,
leading to v0.8 being uninstallable if there has been a point release
(v0.8.1).
2. Change Spack CI to use latest release _and_ develop branch, because
the former is how Arbor is rolled out in various centers, and thus
should be verified to work, and the latter lets us know of upcoming
issues with dependencies.
3. Changed Spack CI to not use `build_spack_package.sh`. This script
does everything in `/tmp`, and Github Actions appear to be wiping `/tmp`
between steps, which means breaking up the run into steps is difficult.
Breaking up is desirable, because then one can see whether building or
testing breaks.
4. Run Python tests after building the Spack package.
5. Also use Spack-provided deps when Spack is instructed to build
master, not submodules.
6. Tiny fix for the `ciwheelbuild` failure.

## Future issues

Out of scope for now, but identified as problems:

1. Currently, Spack cannot execute the unit tests, because the
Google-Test package isn't built and Arbor doesn't look for it. @boeschf
is looking at the latter: making it possible to build with an external
Google Test.
    - Will be solved by https://github.com/arbor-sim/arbor/issues/1731 
2. The former is another problem: Spack does not or not always (it's not
clear) concretize test dependencies. This is an open issue with low
priority: https://github.com/spack/spack/issues/29447
This mean that despite oru `package.py` properly including
`depends_on("googletest@1.12.1", type="test", when="@0.7.1:")`, Spack
does not build it, at least not in the dev-build workflow as we
currently do. And it isn't clear what an always working workaround is,
see issue. (Spack can't display the deps for the `test` variant:
https://github.com/spack/spack/issues/23586 `spack spec -I arbor`
doesn't show Google Test, and neither does `spack dependencies --deptype
test arbor`.)
    - Blocked by upstream.