Allow taking derivatives of `Xd` (#2221)
# Introduction 

The core problem is that `Xd` is not a `STATE` and only those can appear
in derivatives/ODEs.

This requires rewriting the idiomatic
```
 NEURON {
    SUFFIX decay
    USEION x WRITE xd
    RANGE tau
}

PARAMETER { tau = 5 }

BREAKPOINT {
   SOLVE dX METHOD cnexp
}

DERIVATIVE dX {
   xd' = -tau*xd
}
```
as this
```
NEURON {
    SUFFIX decay
    USEION x WRITE xd
    RANGE tau
}

STATE { F }

INITIAL { F = xd }

PARAMETER { tau = 5 }

BREAKPOINT {
   SOLVE dX METHOD cnexp
   xd = F
}

DERIVATIVE dX {
   F = xd
   F' = -tau*F
}
```

_However_, `BREAKPOINT` is translated into _two_ C++ calls, not one:
1. compute $F(t)$ from $F'(t)$
2. update `xd` based on `F`

and other things -- like the **solver** -- happen in between. This has
proven
not only ugly, but also unsound, see #2129 and possibly more.

# Fixing the Problem

Simply allow `Xd` to be differentiated. As this is not an ionic
_current_ nothing
bad will happen., right? Right? This could still be awkward when making
mechanisms like
this
```
BREAKPOINT {
   SOLVE dX METHOD cnexp
   xd = xd + 42
}

DERIVATIVE dX {
   xd' = -tau*xd
}
```