Use Labels for Probes instead of Numerical Ids (#2213)
# Overview

This rectifies a longstanding exception from Arbor's scheme of `(gid,
"label")` to address
point-like (or set-of-point-like) entities. Probes are up to now
addressed as `(gid, lid)`, where
`lid` is computed from the position in the return value of
`recipe::get_probes`. This is not
only a more uniform interface, but also a safer one (imagine adding an
additional probe in the
middle).

That is, from now now on, each probes needs to add a stringy tag, eg
```cxx
  auto get_probes(cell_gid_type gid) const override { return {{cable_probe_voltage(...), "Um"}}; }
```
and we consume this as
```cxx
// ....
auto sim = simulation(rec);
sim.add_sampler({0, "Um"}, ...);
```

In addition the optional `tag` field was removed. It has been hardly
ever been used (no examples
and no models in the wild I know of). It has also been hidden from
Python. Thus, the two APIs are
now closer aligned.


# ⚠️ Breaking

- Each probes needs a string tag.
- Numerical ids are no longer allowed.
- (C++) `probe_info::tag` is gone

# TODO
- [x] Update docs.
- [x] Check python examples.
- [x] Check python tests.

# Future tasks
Here are some things I plan to do later on, but might overburden this
PR. The idea is to streamline
the interface presented to the user and easing their cognitive load.

- We should probe POINT processes by label, too.
- We should streamline the sampler interface, currently there are 4
cases
  1. POINT: single item metadata  of specialised struct
  2. POINT cell: vector of these structs.
  3. DENSITY/voltage/current/...: scalar float, `mlocation`
4. DENSITY/voltage/current/... cell: vector of float, vector of `mcable`
  
Of these we could at least merge 1 & 2, potentially 3 & 4, and maybe
1-4.
- Potentially adding pre-canned objects/hooks to dump data in certain
formats (although I'd rather _not_ take additional dependencies, HDF5 or
otherwise).