re-introduce updated CSCS CI (#2402)
unit tests and examples now run again on CSCS premises:
- eiger: 2x AMD EPYC 7742 (multicore)
- todi: 4x GH200 modules (gpu: arbor is compiled with cuda 12.4 and runs
on the H100)

The distributed tests are not enabled as of now (due to current problems
with the gitlab runners). Once this is fixed, I will add another PR to
enable them.

The tests run automatically for every PR to master, but only for
whitelisted users. Manual trigger: make a comment with the content
`cscs-ci run default`.

Other changes:
- use rpaths for pyarb: helps the linker to find pugixml
- mark some more external CMake variables as advanced (from units)